pyproject.toml
[project]
name = "urlshortner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.2",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
